refactor: group creation API and composition modules
