+++ crates/jet3/src/creation/composer/initial_long_values.rs
@@ -23,7 +23,7 @@ pub(crate) fn initial_payload_start(
     root: PageNumber,
     first_create: bool,
 ) -> Result<u64, ComposeError> {
-    let plan = crate::table_schema_plan::plan_table_schema(table, root.get(), first_create)?;
+    let plan = crate::creation::schema_plan::plan_table_schema(table, root.get(), first_create)?;
     Ok(root.get() + plan.appended_page_count())
 }
