+++ crates/jet3/src/creation/relationship_tests.rs
@@ -200,5 +200,5 @@ fn corrupted_written_page_and_wrong_endpoint_fail_publication_check() -> TestRes
     Ok(())
 }
 
-#[path = "create_relationship_rows_tests.rs"]
+#[path = "relationship_rows_tests.rs"]
 mod initial_rows;
